Show all tasks

+++ oci-react-samples/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/repository/SprintUpdateRepository.java
@@ -11,7 +11,7 @@ import javax.transaction.Transactional;
 @Transactional
 @EnableTransactionManagement
 
-public interface SprintUpdateRepository extends JpaRepository<SprintUpdate,Integer> {
+public interface SprintUpdateRepository extends JpaRepository<SprintUpdate, Long> {
 
 
 }
